requirements.txt
Flask
Flask-JWT-Extended
python-dotenv
mysql-connector-python
werkzeug